PinManager: remove unknown sticky bits in response to MovePin request
We used to receive the MovePin request and if we could not find records in the database, we returned error. This patch changes the behavior, we will now remove the unknown sticky bits that are supposedly owned by this PinManager, but have no records. We will not check the results of the operation, as we would not know what to do with the failures and we do not want to block on waiting the reply here, and we do not want to add complex asynchronous processing either.

Signed-off by: Gerd
http://rb.dcache.org/r/1214/