Unittest pinmanagerTest temporary solution
Based on the discussion in the mailinglist about how to deal with this it seam 
to be most resonable to move the test out of our testsuite for now, 
since it is not running there and it is not easy to solve this now. 
On the other hand I think Timur likes it to be still runnable for him locally.
Will now only exclude this test, but let it stay where it is 
- in the hope this will not be an exception forever.

http://rb.dcache.org/r/1216/

Acked by Tigran