PnfsManager: Add access mask to PnfsMessage
This patch allows checking additional access rights before processing a
PnfsMessage. PnfsMessage has received a new property, accessMask, which
is check in PnfsManagerV3 ahead of processing the message.

Eventually this check should be moved into the name space provider. In
particular for Chimera this would make the check much more efficient.
However for now none of the name space providers support this. I
considered adding the check into PermissionHandlerNameSpaceProvider, but
this turned out to be a total mess. Once we decide to move the check
into Chimera, this will be a PnfsManager internal reorganisation.

The access mask is checked in addition to whatever other checks
PnfsManager performs when processing a message. This has to be this way
because the access mask exclusively refers to a single file system
object. To name an example, the ability to delete a path requires
several permissions: you must be able to lookup all path elements, you
must have DELETE_CHILD permission on the parent directory and DELETE
permission on the file itself. Access mask can however only express
rights on one of those objects.

Right now we do not check all access masks. This is because our current
PermissionHandler interface doesn't allow us to check all of them. This
is not an architectural limitation, but for now I suggest we settle for
a subset so we can move forward.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://chaperon.ndgf.org:2288/usageInfo