junit: switch to hsqldb-1.9-rc6
The new 1.9 version of hsqldb contains correct
SQLSTATE codes, which is heavily used in chimera.
(my original patch was for 1.8, but merged into 1.9)


The patch fixes one of the PnfsManager failing
test. The last failing test is the issue between
hsqldb and chimera, but have to be fixed in chimera.

Target: trunk, 1.9.5
Acked-By: Gerd
Patch: http://rb.dcache.org/r/1229/