infoprovider: remove old info-provider xylophone configuration file
The info-provider recently underwent a fairly extensive change to conform to the 
installed capacity document.  Since this was a large change, the old info-provider was 
kept with a different name.  This allows sites to deploy the old info-provider should 
the new one fail to work for them.

The new (installed-capacity compliant) version of the info-provider is now deployed by 
several sites.  There have been no new issues reported for a while, so the old 
info-provider is now redundant.

This patch removes the old info-provider from Trunk.  We can review whether we need to 
continue supporting the old info-provider in 1.9.5 as a separate issue.

Patch is for trunk

Acked-by: Gerd
RB: http://rb.dcache.org/r/1228/