Reverted r13077
Milton 1.4.3 started has limitted support for the Expect HTTP header. 
Enough to reject all requests with an Expect header, but not enough to 
actually make it work.

Curl specifies an expect header when putting files and hence the upgrade 
to Milton 1.4.3 broke PUT via curl.

Reverting to 1.4.2 until we find a solution.