autotools comes to dcap
RB cant display like the empty directory "config" that should be made in the source directory.

Some issues are not addressed in this patch which maybe should. These include;
Getting the version number from SVN not CVS.
Autoguessing the paths for  --with-globus-lib[=DIR] --with-globus-include[=DIR]
Building rpms when  --with-globus-include and --with-globus-lib are not specified.
Library symbolic links having the version number.
removing the ant target for dcap.
Why are #defines on the command line for compilation.

So with these serious issues not resolved why did I not make these changes. Well the ant stuff is needed already, and its a different subject and in my opinion should be in a different patch that I haven't made yet. The #define issue is left to try to minimize the differences in compilation till a later patch. The other issues are all enhancements to what we have currently and I feel are more incremental and I shoudl like to get them in for follow up patches. Other issues will be addressed in a soon to follow (after this is committed is a source code reorganization into a standard GNU source code layout)


signed off by tigran
http://rb.dcache.org/r/1191/