FTP: Restore Kerberos FTP Door functionality, make Kerberos FTP door work with gPlazma Cell and gPlazma module
This patch makes usage of authRecord conditional. This code also moves the code responsible for communication with gPlazma cell into the GssFtp's abstract method
    protected abstract AuthorizationRecord authorize()
            throws AuthorizationException ;
and moves the current implementation, which only works for gsi door into the GsiFtp door, and adds the new implementation that would work for Kerberos into the Kerberos
FTP door.

This patch also removes unnecessary comments and changes the names of the class variables to be follow the style used in AbstractFtpDoor
Signed-off by: Gerd
http://rb.dcache.org/r/1259/