WebDAV: Redirect GETs on directories with missing /
The HTML we generate for the directory listing contains relative path
references. E.g. a listing of http://ftp1.ndgf.org/pnfs/ndgf.org/
contains a link <a href="data/">....</a>. This resolves to 
http://ftp1.ndgf.org/pnfs/ndgf.org/data/ when the too are concatenated.

We run into trouble if the user requested a listing of 
http://ftp1.ndgf.org/pnfs/ndgf.org (without a trailing slash). In that
case RFC 2396 section 5.2 says this about how to resolve the relative 
path-reference:

      a) All but the last segment of the base URI's path component is
         copied to the buffer.  In other words, any characters after the
         last (right-most) slash character, if any, are excluded.

Thus this link resolves to http://ftp1.ndgf.org/pnfs/data/, which isn't
what we want.


This patch solves this my redirecting the GET on the directory resource
with a missing trailing slash to the URI with the trailing slash. Apache
does the same.

Target: trunk
Target: 1.9.6
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1269/