infoprovider: remove old, Java-based info-provider
The current supported versions of dCache support two info-providers: the old Java-based info provider and the 
newer XML-based one (using the info service).  The XML-based info-provider has existed since about 1.9.1, 
designed (as much as possible) as a drop-in replacement for the Java-based one.

As of 1.9.5, the XML-based info-provider publishes according to the current WLCG guidelines whereas the old 
info-provider publishes according to older standards.  Although the old info-provider could allow sites to 
publish nearline accounting information, this would only be possible by editing static LDIF files.  The 
XML-based info-provider allows publishing nearline accounting information from a concise XML format.

The new info-provider is now widely deployed (I'd estimate about 50% of deployed dCache instances are using it) 
and it seems to work well.

This patch removes the old info-provider.  The source directory (modules/infoProvider) is completely removed and 
the ant build files are updated to remove references to infoProvider.  The jclarens external JAR file is also 
removed as it was only referenced from within the info-provider.

Note: I originally thought to update the dCacheSetup in the same patch.  However, having tried twice to update 
the set of BASH scripts and failed, my feeling is that this is a bigger task than can safely be undertaken in 
the same patch.  After talking with Owen, the default deployment of dCache will use the new info-provider, so I 
believe new installations will continue to work.  Additional patches may be needed to update dCacheSetup so that 
they force sites to migrate to the (new) XML-based info-provider.

Patch is for trunk only

Acked-by: Gerd 
RB: http://rb.dcache.org/r/1264/