findbugs: remove the old info-provider jar from class-path
The find-bugs target in ant requires the class path to be defined.

A previous patch to remove the info-provider failed to remove the old info-provider's JAR file 
(infoDynamicSE.jar) from this target's classpath definition.  This results in attempts to run 
findbugs failing with the following exception:

 [findbugs] Exception in thread "main" java.util.zip.ZipException: Error opening 
[...]/build/classes/infoDynamicSE.jar

This patch fixes this by removing infoDynamicSE.jar from the findpath definitions' classpath.

Patch is for trunk only.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1273