infoprovider: remove scripts that invoke old info-provider
The BDII/GIP process of collecting data from information providers is to execute something and accept that 
process' STDOUT output, which is expected to be LDIF.  To facilitate this, it is common to write a small 
script to invoke some data-gathering activity.

The old infoprovider had two such scripts: one that acted as a GIP plugin and the other that acted as a GIP 
provider.  With the removal of the old info-provider, these scripts are no longer serve any useful purpose.

This patch removes these scripts.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1274

Patch is for trunk.