srm: extract space manager and transfer managers from srm domain
In order to run in terracotta configuration, when we run multiple copies of srm, we must still run just a single copy of the space manager. Also having multiple copies of the space managers do not help. Also separation helps with debugging and load distribution and messaging between srm and transfer managers seems to work better when they are in the separate domain. 

Signed-off by: Paul
http://rb.dcache.org/r/1284/