FTP: Restore Kerberos FTP Door functionality
This patch makes usage of authRecord conditional. This code also moves 
the code responsible for communication with gPlazma cell into the 
GssFtp's abstract method

    protected abstract AuthorizationRecord authorize()
            throws AuthorizationException ;

and moves the current implementation, which only works for gsi door into 
the GsiFtp door, and adds the new implementation that would work for 
Kerberos into the Kerberos FTP door.

This patch also removes unnecessary comments and changes the names of 
the class variables to be follow the style used in AbstractFtpDoor

Target: trunk
Target: 1.9.6
Signed-off-by: Timur Perelmutov <timur@fnal.gov>
Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: http://rb.dcache.org/r/1259/