SRM: Use Jetty to host SRM webapp
This patch introduces a domain which embeds SRM inside Jetty. The
immediate benefits are:

- No need for a custom deployment script
- No custom SRM configuration file
- No separate log4j configuration
- No jar files to copy during installation and hence no risk of
  forgetting to copy them
- No custom log file location
- No custom init script
- No need to ping to wake up SRM
- No 'hiding' of connector related exceptions
- No 'late' activation of proper log file formatting
- No hardcoded paths to /opt/d-cache/

The SRM domain becomes just another dCache domain. It starts the space
manager, thread manager and transfer managers like any other cell would
be started. In addition to this, the jsrm cell is started. It hosts a
Jetty instance and instantiates the SRM Storage cell. The Jetty instance
loads the SRM webapp.

To activate the new domain, add jsrm as a service in etc/node_config.
You can activate either srm or jsrm, but you must not run both at the
same time (as they use the same port).

If and when we are happy with Jetty, we may decide to remove Tomcat. At
that point we can start to simplify the Storage class and don't need to
start it as a separate cell: It can become just another cell component
running inside the spring container.

One could possibly eliminate the webapp layer and start the Axis servlet
directly in Jetty. This would eliminate the extra classloader.

The only changes to the SRM is to make loading the log4j configuration
file and the dcache parameter file optional.

Notice that the SRM webapp used for jsrm doesn't contain any jar files. 
All jar files are loaded through the outer classloader.

The name 'jsrm' is temporary. If we should decide to move from Tomcat
to Jetty, then I imagine we will rename jsrm to srm.

The jetty-plus jar must be added to modules/external/jetty/ for this patch 
to work. I will commit that jar together with this patch.

The files in the WEB-INF directories are copied from the current srm-webapp
and slightly modified.

Testing Done:

I ran my usual transfer tests, and they all passed.


Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1295/