infoprovider: update BDII-bug workaround so all RFC-2253 must-markup characters are converted to a colon
BDII contains a bug that is triggered if an object's DN contains characters that must be escaped (see 
RFC 2253).  The current info-provider contains a work-around that maps these characters to the colon 
character.  With current usage, only NAS-derived (installed capacity) GlueSA objects will contain such 
must-markup characters.

The work-around was intended to map all marked-up characters to the colon character; however, the 
back-slash and the equals characters were mistakenly omitted.  The NAS description "decoration" 
characters do not use these characters, but they may still appear in the DN if the store unit name 
contain them.

The fix is to add '=' and '\' to the list of characters that are converted to a colon.

Patch is for all supported branches: Trunk, 1.9.6, 1.9.5, 1.9.4

RB: http://rb.dcache.org/r/1342/
Acked-by: Owen