Reinstate per platform preload calling functions
AC_CANONICAL_SYSTEM is used to detect the platform. Then if platform is known preload is enabled. Each platform can override the default preload functions to call the preload methods.

Benifits:

a) Less rubbish on the command line while compiling.
b) Each platforms preload calls can be specified in a clear way in a single file without too much extras.
c) Since preloads have defaults only the things that differ on a per platform basis need to be specified.
d) Restores the functionality of the per OS Make files in a single configure script. 

Things to consider:

Maybe each of the functions selected for each platform should be tested if they exist.


Acked-by: Tigran
http://rb.dcache.org/r/1291/