PnfsManager: Fix compatibility with 1.9.5 cells
PnfsGetFileMetaDataMessage in 1.9.6 inherits from PnfsGetFileAttributes.
This was designed to be compatible with 1.9.5 cells. The Java
Serialization spec says:

# Removing classes - Comparing the class hierarchy in the stream with
# that of the current class can detect that a class has been deleted. In
# this case, the fields and objects corresponding to that class are read
# from the stream. Primitive fields are discarded, but the objects
# referenced by the deleted class are created, since they may be
# referred to later in the stream. They will be garbage-collected when
# the stream is garbage-collected or reset.

Thus one can deserialize a PnfsGetFileMetaDataMessage from 1.9.6 in
1.9.5, because 1.9.5 is simply supposed to discard the extra data in the
intermediate PnfsGetFileAttributes class in the class hierarchy. I took
care that the old fields used by 1.9.5 are still set to the correct
values.

We however discovered the bad way that compatibility did in fact break.
The crucial bit is that "the objects referenced by the deleted class are
created". That means even thought 1.9.5 will throw away the data in the
fields of the new intermediate class, it will still deserialize the
field values. Hence it will deserialize the FileAttributes object
referenced in the PnfsGetFileAttributes. 1.9.5 knows the FileAttributes
class, so that's not an issue. It does however not know about the
ACCESS_TIME enum value of FileAttribute referenced in the FileAttributes
object. As PnfsGetFileMetaDataMessage uses this enum value,
deserialization failed on 1.9.5. This broke writing to tape at NDGF.


This patch adds another temporary workaround to the existing
compatibility workarounds in PnfsGetFileMetaDataMessage. We can detect
that the request comes from 1.9.5 by recognizing that the _attributes
field is null. The patch modifies the setFileAttributes method such that
we do not set the FileAttributes property if _attributes is null. This
is okay because the client would discard it anyway. This solves the
compatibility issue because the 1.9.5 cell will not be forced to
deserialize this field.

All these workarounds can be discarded once we decide to break
compatibility with 1.9.5 (likely in 1.9.7 or 1.9.8).

Target: trunk
Target: 1.9.6
Patch: http://rb.dcache.org/r/1347/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>