jms: Listen to wildcard address rather than localhost
With the upgrade to ActiveMQ 5.3, the embedded broker would only listen
to the loopback interface, not the wildcard interface as it did before.
This is due to https://issues.apache.org/activemq/browse/AMQ-2094.

This patch updates the configuration of the embedded broker to listen on
the wildcard address.

Eventually we should make the setting configurable, but for now I just
want to restore the old behaviour.

Target: trunk
Acked-by: Jan Schaefer
Patch: http://rb.dcache.org/r/1349/