xylophone: tidy-up of DN calculation and link-breaking.
The xylophone package provides a generic framework for converting XML to 
LDIF.

The following patch synchronises the version in dCache with the external 
repository.  The changes include:
  1. removal of one empty line in the preamble,
  2. DN for an object is calculated using a more robust method: passing the 
     parent object's DN to the child object. (the method used in dCache's 
     xylophone currently is to iterate backwards for each object in the hierachy, 
     calculate the RDN values),
  3. End-of-line character handled when undertaking LDIF line-breaking is 
     re-factored slightly,
  4. Some minor changes, such as making template names more descriptive, 
     updating some comments, etc.

Apart from the removal of a single empty line (which is ignored by a 
conforming LDIF parser), the output from running the info-provider after 
applying the patch will be identical.

This work is preparation for including some additional functionality that 
will allow us to publish the dCache versions of all doors and pools used by 
a site.

Patch is for all supported branches: Trunk, 1.9.6, 1.9.5

RB: http://rb.dcache.org/r/1339/
Acked-by: Owen