BUG FIX : ./configure with no prefix set the prefix to NONE for install docs.
Since we had to back port to support old versions of autoconf as in RHEL and SL5 we had to set our own prefix handling.

This was all fine but since autoconf defines ${prefix} as NONE if not set and it then gets evaluated later, automake used the templated value not the evaluated value of "/usr/local", all worked as expected when --prefix was set.

This fixes that bug.

Signed-off by: Paul
http://rb.dcache.org/r/1371