srm: fix startup-ping so failures are reported.
The SRM (in fact, the Axis servlet) requires a request to initialise the SRM
components.  There is a small, custom SRM-Ping client code-base that will undertake
to ping the SRM, so starting the dCache domain.

Currently, should this SRM-Ping client fail, no error message is reported.  This is
unfortunate as the site-admin will otherwise have no knowledge of a failing
SRM-startup.  Although the next request the SRM receives should start the SRM, any
failures should be reported if for no other reason than the problem may be shared
between the srm-ping client and the SRM server itself.

This patch adds support for reporting any errors during the SRM startup.

RB: http://rb.dcache.org/r/1377/
Acked-by: Dmitry
Acked-by: Owen