cells: Make cellClass configurable
The primary purpose of the cell class field of CellInfo is to provide
information about what kind of cell this is. The main user of this
information is the dCache GUI. The cell class field used to be the short
name of the Java class implementing the cell, however with the
introduction of UniversalSpringCell the Java class no longer uniquely
identifies the kind of cell.

This patch makes the cell class configurable and defines a custom cell
class for all UniversalSpringCells.

As this is essential for the GUI to work well, I propose we also apply
this patch to 1.9.5 and 1.9.6.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1366/