config: Fix configuration parameter override in batch files
Turns out that overriding configuration parameters on the exec line of
the batch files didn't work at all. The problem is that the arguments
that override configuration parameters where loaded through another
exec, and that exec had an empty argument list. Hence nothing was
imported.

To resolve the issue I had to add the argument list to the exec
statement calling the import-arguments.fragment. That has the nice side
effect of collapsing the two variants of the fragment into one.

Target: 1.9.6
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1395/