scripts: Fix handling of multiple dcap doors
There was an architectural flaw in the init script regarding handling of
multiple DCAP door instances. The problem was that configuration files
would be loaded per service type, not per domain. That is correct for
pools (which share a common poolSetup file), but not correct for DCAP
doors (which have a Setup file per door instance).

This patch updates the scripts such that a configuration file can be
loaded for a service or for a domain. This covers both use cases.

Notice that the install.sh script has a flaw regarding how it extracts
the dcapPort. It assumes there is one correct port, however this is not
the case if multiple DCAP doors exist. This patch does not address that
issue.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Target: 1.9.4
Ticket: #5330
Acked-by: Owen Synge <owen.synge@desy.de>
Patch: http://rb.dcache.org/r/1344/