cells: Refactor CommandExitException handling
To ensure that the source of errors is included when existing, this
patch refactors the CommandExitException handling. Rather than
processing it in CellShell.objectCommand2, we process it in
CellShell.execute, which is the only place that utilized onerror setting
anyway.

We also make certain that line numbers are included in IOExceptions and
RuntimExceptions.

Testing Done:

Added the following to dCache.batch: 

create diskCacheV111.srm.dcache.Storage test8 "-jdbcUrl=sds -jdbcDriver=sds -dbUser=srmdcache -dbPass=test -pgPass=/tmp/x"

After starting dCacheDomain the log file contained: 

14 Jan 2010 13:59:26 (batch) [] could not get password from '/tmp/x' for  hostname: 'localhost' ,port: 5432 ,database: 'sds' and username: 'srmdcache' 
14 Jan 2010 13:59:26 (batch) [] (666) /opt/d-cache/config/dCache.batch: line 12: (3) java.sql.SQLException: could not get password from '/tmp/x' for  hostname: 'localhost' ,port: 5432 ,database: 'sds' and username: 'srmdcache'  from ac_create_$_2_3; caused by: (3) java.sql.SQLException: could not get password from '/tmp/x' for  hostname: 'localhost' ,port: 5432 ,database: 'sds' and username: 'srmdcache'  from ac_create_$_2_3; caused by: could not get password from '/tmp/x' for  hostname: 'localhost' ,port: 5432 ,database: 'sds' and username: 'srmdcache' 

The first log entry is generated by the SRM (it unfortunately logs and 
throws an exception; such code will typically result in multiple errors 
being logged) and the second by the batch interpreter when it is 
exiting. Notice that no stack trace is printed as the exception is not a 
RuntimeException.

Target: 1.9.5
Target: 1.9.6
Target: trunk
Ticket: #5329
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1393/