pool: Fix migration of deleted files
The migration module for obvious reasons fails to migrate files that
have been deleted in the namespace. It would however not recognize this
as a permanent error and retry the file indefinitely (or at least until
the file had been garbage collected).

This patch introduces the possibility for migration tasks to fail with a
permanent error. Such errors are logged, the task is not retried, and
the job does not sleep.

I had to fix error generation for FileNotFoundCacheException in the
Chimera name space provider: The error message was rather inconsistent
and this would cause the error logged by the migration module to contain
an empty string.

This problem has been reported as a bug by Mattias Wadenstein.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1420/