scripts: Fix 'dcache stop' behaviour in case of domain restart
Commit r12452 broke stopping domains in the process of auto restarting.
Or rather, the commit uncovered a bug that was hidden by the fact that
we would stop a domain no matter whether we though it was running or
not. The problem is that our PID files store the PID of the java
process. However during an auto restart there is no java process, and as
a consequence our scripts believe the domain is not running. Since we
pause for 10 seconds during an auto restart, this situation is not as
unlikely as it may sound.

The fix is rather elaborate. The daemon script is updated to support two
PID files: One for the wrapper script and one for the child (the java
process). The first will be used to determine if the domain is running,
the second to determine the PID for dumping the heap and threads.
generic.lib.sh is updated to support these two PID files, as is the
dcache script and the shell libraries.

At the same time I made the daemon script and the wrapper2.sh script use
'set -e'. Thus any unchecked command with a non-zero error code will
cause the scripts to exit. This is a great way to ensure fail-fast
behaviour, but it also required changes the code a number of places that
ran unchecked commands followed by evaluating $?. Those places have been
updated to evaluate the return code right away, either in an if or
through || or && after the command.

Notice that Solaris sh doesn't allow negating the return code of a
pipeline (i.e. you cannot use ! in front of a command). Thus we have to
use || rather that 'if !'.

Due to the 'set -e' changes, the patch is maybe too big for the stable
branches. If the reviewer finds those changes to be too risky for the
stable branches, then I will separate the PID file changes from the
error handling. Alternatively part of r12452 may be reverted on stable
branches and this patch is only committed to trunk.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Ticket: #5409
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1417/