WebDAV: Fix address resolution for PUT
PUT is currently proxied through the WebDAV door. The pool opens a TCP
connection to the door and the door relays the data to the pool. For
this to work the door needs to send its IP address to the pool. This
didn't work as the WebDAV would send the wildcard address (0.0.0.0) to
the pool rather than a public IP address of the door. This worked fine
on my single node test installation, but failed on real installations.

This patch addresses this problem in two ways: First of all it uses
InetAddres.getLocalHost() to obtain a real address of the door. Second,
it adds the configuration parameters webdavAddress and
webdavInternalAddress. The former defines the IP address to listen to
for connections from WebDAV clients. The latter defines the IP address
to listen to for connections from the pool.

LoginBrokerHandler was updated to allow for the door address to be set
as a String from Spring.

Target: trunk
Target: 1.9.6
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1409/