WebDAV: Refine error message reporting to billing
We know that CacheException always got a proper message string. Hence we
use getMessage() to get the message string to send to billing.

RuntimExceptions and IOExceptions on the other hand often embed the
failure in the name of the exception and include no error message. Using
toString() on those is essential to get a meaning full message (e.g.
NullPointerException or AsynchronousCloseException don't have an error
message).

A race condition in relayData() was fixed: If a mover would report back
an error rather than connecting to the door, then the server channel is
closed in the DoorTransferFinishedMessage processing. This will ensure
that the accept() call in relayData() is interrupted. However the
AsynchronousCloseException doesn't provide much information about what
went wrong. Also, if the DoorTransferFinishedMessage would be received
before relayData() calls accept, then the server channel would be null
and we got an NPE. The patch changes the semantics of relayData() such
that it waits for the mover to terminate and rethrows whatever failure
the mover produced.

join() is renamed to waitForMover() to improve the readability of the
code.

Target: 1.9.6
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1406/