dcap: library version names + extras
The new autotools version of the dcap Makefiles does not propagate the library version to the library filenames the way the old plain Makefiles did. This patch fixes this issue.

It also fixes some issues with undefined symbols due to missing libraries in the linking.


Committing on behalf of: Mattias Ellert
Signed-off by: Owen
http://rb.dcache.org/r/1413/