WebDAV: Fix error handling in filters
MiltonHandler would commit the response in a finally block. Thus even
when some of the filters would throw a RuntimException, the response
back to the client would be 200 OK.

This patch removes the finally block allowing Jetty to handle the
exception and generate a 500 Server Error.

LoggingFilter was updated to extract user information directly from the
request object. Thus the filter can be placed in front of the
SecurityFilter and hence will log failures that happen during
authentication.

Target: trunk
Target: 1.9.6
Ticket: #5416
Patch: http://rb.dcache.org/r/1444/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>