pool: Fix stage error handling
If the stage script failed with an error code, the pool would still
attempt a checksum verification and it would still commit the
WriteHandle. This patch fixes the error handling such that the operation
is aborted with an exception.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Target: 1.9.4
Ticket: #5408
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1437/