pool: Don't disable pool on sync
In r13191 we changed the handling of sync failures such that rather than
just logging a warning we disable the pool. Since then we have found
that sync may fail in certain non-critical situations, for instance with
network file systems.

This patch changes the behaviour again such that a sync failure does not
disable a pool. Instead the sync is repeated and if the second sync
fails too, then the transfer is failed. The motivation for repeating the
sync is that an fsync() call can fail due to it being interrupted by a
unix signal. Such a failure is transient.

Target: trunk
Target: 1.9.6
Target: 1.9.5