gridftp: Fix interpretation of QUIT command
RFC 959 says:

         LOGOUT (QUIT)

            This command terminates a USER and if file transfer is not
            in progress, the server closes the control connection.  If
            file transfer is in progress, the connection will remain
            open for result response and the server will then close it.
            If the user-process is transferring files for several USERs
            but does not wish to close and then reopen connections for
            each, then the REIN command should be used instead of QUIT.

            An unexpected close on the control connection will cause the
            server to take the effective action of an abort (ABOR) and a
            logout (QUIT).


And later it says:

   Certain commands (such as ABOR, STAT, QUIT) may be sent over the
   control connection while a data transfer is in progress.

Hence we should honor the QUIT command even while a transfer is in
progress. We should however not abort the transfer. Instead we should
wait for the transfer to complete and then we close the control channel.
If the control channel is closed by the client, then we are supposed to
abort the transfer.

Although technically a bug fix, I only target the patch at trunk as
changing the interpretation of the FTP protocol could potentially
introduce compatibility issues with some clients.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1438/