pool: Refactor and fix state handling for volatile pools
Since 1.7.0 we have had an issue with how we handle lfs=volatile: If
access latency was ONLINE, then the file would receive a System sticky
bit even though the pool was marked volatile.

This patch refactors the definition of the state of a newly uploaded
replica: The Strategy pattern is used to decouple the algorithm for
deciding which target state or sticky bits to apply to an implementation
of the ReplicaStatePolicy interface. VolatileReplicaStatePolicy is used
on volatile pools and ALRPReplicaStatePolicy is used on all other pools.

LFSReplicaStatePolicyFactory contains a static factory method for
creating a ReplicaStatePolicy based upon the lfs mode. I thought I could
used Spring SpEL for this, but SpEL expression cannot be used in the
class attribute for the bean tag.

CacheRepositoryV5 is changed to no longer alter the target state on
volatile pools. This means that it now respects whatever target state is
requested by the caller. This alters the behaviour of volatile pools
when used as targets for migration tasks and other cases where a file is
written to a pool from other sources than a client upload. The
ReplicaStatePolicy is only consulted for files uploaded from a client or
when recovering pool meta data.

The LFS mode configuration option mapping is moved to the Spring XML
file. Unfortunately we still got a p2p mode. This is inparticular a
problem for installations with a hopping manager. Files moved by the
hopping manager should be subject to the same ReplicaStatePolicy, but at
the moment the state is a result of a combination of the p2p request,
the lfs mode and the p2p mode.

The admin command to change the LFS mode on the fly has been removed. It
doesn't really make sense to change the LFS mode of an existing pool and
the option was also documented as being for debugging purposes only.

Target: trunk
Ticket: #5325
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1432/