info: switch state read/write lock to use the fair variant.
The info-service maintains an overview of information about dCache.  It is updated with fresh 
information with a schedule controlled by the info-service itself.  As fresh information is 
received from remote components, requests to update the state are enqueued.  A single writer 
thread is responsible for updating dCache state by taking one update after the next from the 
queue and processing them.

Concurrently the info-service may receive requests for some (or all) of dCache's current state.  
These requests will (typically) come from the web interface, but there are other potential 
sources of these requests.  Without any locking mechanism, there would be the risk that a read 
request would be satisfied whilst a write request is being processed.  This would risk returning 
an inconsistent result-set, or triggering concurrent read/modification of a tree-node.

To eliminate this risk, a read/write lock is employed.  The current read/write lock is the 
default (no-constructor-argument) ReentrantReadWriteLock from the java.util.concurrent.locks 
package.  Lock objects of this type are unfair: the JavaDoc describe this unfairness that "a 
nonfair lock that is continously contended may indefinitely postpone one or more reader or writer 
threads".  This behaviour has been observed at US-CMS where (under load) a request to read the 
file is not satisfied for many [some O(10)] seconds.

This patch switches the read/write lock to use the "fair" implementation.  With this, the longest 
waiting Thread will be served next, whether that be a request for a read- or write- lock.  Since, 
in general there are two threads in operation (the update thread and the message thread), this 
means that a read request will block for a maximum of the time taken to process an write 
operation.

Patch is for 1.9.5, 1.9.6 and trunk.
svn ci modules/dCache/org/dcache/services/info/base/State.java
RB: http://rb.dcache.org/r/1408/

Acked-by: Tigran