admin: Fix binary mode (aka GUI support)
The GUI sends $BINARY$ as a string to the admin door to initiate binary
mode. It is expected that the door replies back with the same string.
The code for doing so was present, but we forgot to flush the output
buffer.

Target: trunk
Target: 1.9.6
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1491/